// src/app/core/modules/openapi/model/login-request.ts
/**
 * Adres API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface LoginRequest { 
    username: string;
    password: string;
    email: string;
    phone?: string;
}